Add files via upload
Local’s Moran I showed us that there was a spatial autocorrelation between the parameters, whereas the Local Getis-Ord G mapped the hot and cold spots with confidence level associated to it. The counties in the Local Moran’s I are different than those with Getis-Ord because the confidence level changes such as 95%, 99% with hot and cold spots in Getis-Org G, but with Local Moran’s I there is no distinction in terms of significance level and if there is a hotspot it is shown in the same color for all significance levels, unless it is not significant.